move notifications under routing/common/notifications
